Move state formatting code to format.rs
